moved file to test folder moved json encoding/decoding of db configuration in io module removed unnecessary test_postgres module
