+++ packages/simcore-sdk/tests/nodeports/connection_config.json
